refactor: split out the offer-compiler to a separate file
